Expanded test cases to cover samsum, gsm8k, alpaca and imdb dataset

+++ tests/finetune/test_finetune.py
@@ -16,25 +16,100 @@ from torch.utils.data import DataLoader
 import QEfficient
 import QEfficient.cloud.finetune
 from QEfficient.cloud.finetune import main as finetune
+from pathlib import Path
+import requests
+
+alpaca_json_path = Path.cwd() / "alpaca_data.json"
 
 
 def clean_up(path):
-    if os.path.exists(path):
+    if os.path.isdir(path) and os.path.exists(path):
         shutil.rmtree(path)
+    if os.path.isfile(path):
+        os.remove(path)
+
+
+def download_alpaca():
+    alpaca_url = "https://raw.githubusercontent.com/tatsu-lab/stanford_alpaca/refs/heads/main/alpaca_data.json"
+    response = requests.get(alpaca_url)
+
+    with open(alpaca_json_path, "wb") as f:
+        f.write(response.content)
 
 
 configs = [
     pytest.param(
         "meta-llama/Llama-3.2-1B",  # model_name
+        "generation",  # task_type
         10,  # max_eval_step
         20,  # max_train_step
+        "samsum_dataset",  # dataset_name
+        None,  # data_path
         1,  # intermediate_step_save
         None,  # context_length
         True,  # run_validation
         True,  # use_peft
         "qaic",  # device
+        0.00232291,  # expected_train_loss
+        1.00232565,  # expected_train_metric
+        0.02060609,  # expected_eval_loss
+        1.02081990,  # expected_eval_metric
         id="llama_config",  # config name
-    )
+    ),
+    pytest.param(
+        "meta-llama/Llama-3.2-1B",  # model_name
+        "generation",  # task_type
+        10,  # max_eval_step
+        20,  # max_train_step
+        "gsm8k_dataset",  # dataset_name
+        None,  # data_path
+        1,  # intermediate_step_save
+        None,  # context_length
+        True,  # run_validation
+        True,  # use_peft
+        "qaic",  # device
+        0.0043353,  # expected_train_loss
+        1.0043447,  # expected_train_metric
+        0.0117334,  # expected_eval_loss
+        1.0118025,  # expected_eval_metric
+        id="llama_config",  # config name
+    ),
+    pytest.param(
+        "meta-llama/Llama-3.2-1B",  # model_name
+        "generation",  # task_type
+        10,  # max_eval_step
+        20,  # max_train_step
+        "alpaca_dataset",  # dataset_name
+        alpaca_json_path,  # data_path
+        1,  # intermediate_step_save
+        None,  # context_length
+        True,  # run_validation
+        True,  # use_peft
+        "qaic",  # device
+        0.0006099,  # expected_train_loss
+        1.0006101,  # expected_train_metric
+        0.0065296,  # expected_eval_loss
+        1.0065510,  # expected_eval_metric
+        id="llama_config",  # config name
+    ),
+    pytest.param(
+        "google-bert/bert-base-uncased",  # model_name
+        "seq_classification",  # task_type
+        10,  # max_eval_step
+        20,  # max_train_step
+        "imdb_dataset",  # dataset_name
+        None,  # data_path
+        1,  # intermediate_step_save
+        None,  # context_length
+        True,  # run_validation
+        False,  # use_peft
+        "qaic",  # device
+        0.00052981,  # expected_train_loss
+        0.55554199,  # expected_train_metric
+        0.00738618,  # expected_eval_loss
+        0.70825195,  # expected_eval_metric
+        id="bert_config",  # config name
+    ),
 ]
 
 
@@ -42,18 +117,25 @@ configs = [
 @pytest.mark.on_qaic
 @pytest.mark.finetune
 @pytest.mark.parametrize(
-    "model_name,max_eval_step,max_train_step,intermediate_step_save,context_length,run_validation,use_peft,device",
+    "model_name,task_type,max_eval_step,max_train_step,dataset_name,data_path,intermediate_step_save,context_length,run_validation,use_peft,device,expected_train_loss,expected_train_metric,expected_eval_loss,expected_eval_metric",
     configs,
 )
-def test_finetune(
+def test_finetune_llama(
     model_name,
+    task_type,
     max_eval_step,
     max_train_step,
+    dataset_name,
+    data_path,
     intermediate_step_save,
     context_length,
     run_validation,
     use_peft,
     device,
+    expected_train_loss,
+    expected_train_metric,
+    expected_eval_loss,
+    expected_eval_metric,
     mocker,
 ):
     train_config_spy = mocker.spy(QEfficient.cloud.finetune, "TrainConfig")
@@ -69,8 +151,11 @@ def test_finetune(
 
     kwargs = {
         "model_name": model_name,
+        "task_type": task_type,
         "max_eval_step": max_eval_step,
         "max_train_step": max_train_step,
+        "dataset": dataset_name,
+        "data_path": data_path,
         "intermediate_step_save": intermediate_step_save,
         "context_length": context_length,
         "run_validation": run_validation,
@@ -78,16 +163,20 @@ def test_finetune(
         "device": device,
     }
 
+    if dataset_name == "alpaca_dataset":
+        download_alpaca()
+
     results = finetune(**kwargs)
-    assert np.allclose(results["avg_train_loss"], 0.00232327, atol=1e-5), "Train loss is not matching."
-    assert np.allclose(results["avg_train_metric"], 1.002326, atol=1e-5), "Train metric is not matching."
-    assert np.allclose(results["avg_eval_loss"], 0.0206124, atol=1e-5), "Eval loss is not matching."
-    assert np.allclose(results["avg_eval_metric"], 1.020826, atol=1e-5), "Eval metric is not matching."
+    assert np.allclose(results["avg_train_loss"], expected_train_loss, atol=1e-5), "Train loss is not matching."
+    assert np.allclose(results["avg_train_metric"], expected_train_metric, atol=1e-5), "Train metric is not matching."
+    assert np.allclose(results["avg_eval_loss"], expected_eval_loss, atol=1e-5), "Eval loss is not matching."
+    assert np.allclose(results["avg_eval_metric"], expected_eval_metric, atol=1e-5), "Eval metric is not matching."
     assert results["avg_epoch_time"] < 60, "Training should complete within 60 seconds."
 
     train_config_spy.assert_called_once()
     generate_dataset_config_spy.assert_called_once()
-    generate_peft_config_spy.assert_called_once()
+    if task_type == "generation":
+        generate_peft_config_spy.assert_called_once()
     get_custom_data_collator_spy.assert_called_once()
     get_longest_seq_length_spy.assert_called_once()
     print_model_size_spy.assert_called_once()
@@ -122,12 +211,19 @@ def test_finetune(
         f"{train_config.gradient_accumulation_steps} which is gradient accumulation steps."
     )
 
-    saved_file = os.path.join(train_config.output_dir, "complete_epoch_1/adapter_model.safetensors")
+    if use_peft:
+        saved_file = os.path.join(train_config.output_dir, "complete_epoch_1/adapter_model.safetensors")
+    else:
+        saved_file = os.path.join(train_config.output_dir, "complete_epoch_1/model.safetensors")
     assert os.path.isfile(saved_file)
 
     clean_up(train_config.output_dir)
     clean_up("runs")
+    clean_up("qaic-dumps")
     clean_up(train_config.dump_root_dir)
+
+    if dataset_name == "alpaca_dataset":
+        clean_up(alpaca_json_path)
 
 
 # TODO (Meet): Add seperate tests for BERT FT and LLama FT
